requirements.txt
Django==2.1
mysqlclient==1.3.12
python-decouple==3.0
xhtml2pdf==0.2.3
python-dateutil==2.8.0
django-debug-toolbar==2.0
django-easy-audit
whitenoise==5.0.1
